+++ HospitalManagement/src/main/java/com/hospital/repository/EmployeeRepository.java
@@ -17,7 +17,7 @@ public interface EmployeeRepository extends JpaRepository<Employee, Long>, Emplo
     @Query("SELECT et FROM ExaminationType et")
     public Set<ExaminationType> findAllExamination();
     @Query("SELECT e FROM Employee e WHERE e.specialityId = :id")
-    public List<Employee> findEmployeesBySpeciality(@Param("id") int speciality_id);
+    public List<Employee> findEmployeesBySpeciality(@Param("id") Long speciality_id);
     @Query("SELECT e FROM ExaminationType et LEFT JOIN Employee e ON et.specialityId = e.specialityId" +
             " WHERE et.id = :id")
     public List<Employee> findExaminatorsByExaminationType(@Param("id") long examination_type_id);
